requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
ecdsa==0.14.1
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
python-jose==3.1.0
pytz==2019.3
rsa==4.0
six==1.13.0
sqlparse==0.3.0
